requirements.txt
requests
beautifulsoup4
Flask
python-telegram-bot
